refactor!: move api hooks methods to api module
