+++ tests/test_open_feature_client.py
@@ -2,10 +2,10 @@ from unittest.mock import MagicMock
 
 import pytest
 
+from open_feature.open_feature_api import clear_api_hooks, add_api_hooks
 from open_feature.exception.error_code import ErrorCode
 from open_feature.exception.exceptions import OpenFeatureError
 from open_feature.flag_evaluation.reason import Reason
-from open_feature.hooks import clear_api_hooks, add_api_hooks
 from open_feature.hooks.hook import Hook
 from open_feature.open_feature_client import OpenFeatureClient
 from open_feature.provider.no_op_provider import NoOpProvider
